InvitationRedeemer
This extracts out methods from the both InvitationRedeemers into a
parent class.

This commit sets up `GroupAssignmentInvitationRedeemer` to match the
changes from #58.

`#find_group` now verifies that the user hasn’t already joined a group
if they are trying to create a new one. If they have it just returns
the group they belong to.