Use ActiveRecord callbacks for model creation
When you create a model on Classroom, it will fire of an API call to
GitHub.com in order to create the corresponding object on the site
(Team, Repo etc). Additionally if you remove the object from the
database (a RepoAccess) for example, it will also deletes the GitHub.com
representation (the Team in this case).

This makes the code easier to follow and allows for re usability. The
creation of teams and repositories has been extracted out to Rails model
concerns so that any model can have team creation or repository creation
without too much work.