Invite users to Classroom by invitation email
This will invite a new user by email to join the GitHub classroom. It
creates a new User model on the database in a ‘pending’ state. After
they sign in, their access token is set and they become ‘active’.

Currently you can only invite members who are already admins of the
GitHub Organization. This will be changed to invite anyone in the
future.

Previously we checked almost every request to see if the user was an
admin of the Organization. This is really slow, instead a route
constraint has also been added so that only members of the organization
and staff of classroom can see anything about the
organization including assignments etc.