Require Confirmation for accepting AssignmentInvitations
When the user went to the page it automatically added them to the org
without a confirmation. This changes that so that the user has to click
on the accept button.

The styling still needs some :heart:, but the functionality is there.

I have also changed it so that when the user creates their personal team
instead of it being "Team 1" it just sets it to be there GitHub Login.