Move starter code cloning and pushing to a background job
If the user has a really large repo, the application has the potential
to time out while performing the action.

This moves the process out to a background job on it's own queue
`starter_code` which has a higher weight than the default queue so that
the assignments will be cloned and pushed faster.