Gracefully handle invalid starter code input
Originally from #299, if the user inputs a starter code repo in an
invalid format, in the case of the issue the whole GitHub URL, the app
returned a 500 error.

Now if the format is invalid or GitHub returns that the format is valid,
we will redirect back to the new page and display a nice flash message.