stop modifying org memberships unnecessarily
If the user is already a member of the organization then we should not
try to add them as a member again.

Performing this action causes issues such as #332.

We should also not remove the users membership to the organization if
they are an owner, an example of this would be #337.