merge rails 5 puma config
classroom's settings were copied ontop of the config/puma.rb file for a
rails 5 app. Behavior will stay exactly as it was except that the new
rails standard RAILS_MAX_THREADS env var will be used if the existing
MAX_THREADS env var is not set. This should result in no change because
the MAX_THREADS env var will be set. When running the rails 5 update
script, simply do not overwrite.