Refactor GitHubResource logic
Instead of grabbing from the cache a bunch of times just make the call
once and set all of the attributes at the same time.

If we want to get an attribute with the no cache header we now have
convenience methods `login_no_cache` etc to do that for us.