This one is a doozy
There are two acts to this play:

Act 1: Record loading

Currently we are trying to load every `User` record that is related
to an `Assignment`, `RepoAccess`, or `RosterEntry` and then try iterate
over all of them. Instead we now only select ids and load the `User` for
the current paginated page of results.

Act 2: An unholy amount of API calls, cached or uncached

We were iterating over all of the unlinked users several time and
making API calls for each user for every unlinked `RosterEntry` sort of
like this:

```ruby
roster.roster_entries.each do |roster_entry|
  unlinked_users.each do |user|
    puts "Link #{user.github_user.login} to #{roster_entry.id}"
  end
end
```

For larger classrooms this could result in thousand's of cached or uncached
calls to the GitHub API.

Instead of hitting the API we're going to use ElasticSearch and
grab all of the records.

Why you ask?

Because the `StafftoolsIndex::User` type already has the GitHub login for that user. Saving us tons of time.

Eventually we should move over so that the User has it's own index, but
Stafftools is good enough for now.