Fix tabs not restoring selected tab after refresh
This bug caused the roster tabs on the management page
to forget the currently selected tab on page refresh.
In order to fix this, the roster_tabs js file
has been replaced with a more generic manage-tabs
implementation, which both fixes the bug and makes the addition
of more tabs a little easier. All groups of tabs have been updated
to use manage_tabs.