User `Octokit::Client::Authorizations#check_application_authorization` to check scopes
Previously if you did not pass in a client we would make API calls such as:

```
https://api.github.com/user?client_id=foo&client_secret=bar
```

This would never work, I'm not 100% how we're swallowing up the error.

This updates the method to always use https://developer.github.com/v3/oauth_authorizations/#check-an-authorization

So that we can cleanup the method and not run afoul of of normal rate limits for failed login attempts.