Fix destroy_resource_job_spec
Rails 5.2.x introduced a change to `assert_performed_with` that caused
`destroy_resource_job_spec` to start failing:
https://github.com/rails/rails/pull/33850/files

Here is how the spec was failing:
```
ActiveJob::DeserializationError:
Error while trying to deserialize arguments: Couldn't find Organization with 'id'=1
```

Note that a side effect of the above PR is that job arguments are now
deserialized via `ActiveJob::Arguments.deserialize`.

So what was happening was:
- the block to assert_performed_with was executed, which deletes the
organization
- then in order to check the assertion, the enqueued job's arguments are
deserialized once again, at which point ActiveJob tries to load the
now-deleted Organization, which blows up

I took a step back and asked "what does this test actually care about?".
Based on the "it destroys the resource" description, I re-worked the
test to assert that the associated resource is deleted from the databse
when the job is run.