Fixes Uninitialized constant error when someone visits api/internal/classroom route
We don't have a controller for Organizations in the API namespace, but since we use `resources :organizations`
in the routes, rails expects a controller to be present for the route.
In order to avoid this problem, we can either add a new Organizations controller with a index action
or we can just replace resources with the scope option. Since we don't have immediate plans to add
organizations API, using the scopes options seems a good fit for the fix.