Swallows GitHub::Error when silently deleting repo access
We already send any github API to failbot, so we can swallow this
error. This fixes the MembershipEventJob going into retry cycle in
case the access to the repo was manually removed and hence deleting
repo_access results in `GitHub::Forbidden` error. This can also
occur in groups_controller if repo_access was manually deleted
and someone tries to remove membership on classroom before the
webhook gets triggered/processed.