Fix GroupAssignmentRepo related specs
After we moved from using `ActiveRecord::Callbacks` to a `Creator`
for creating `GroupAssignmentRepo` testing it became much easier.
With this commit, we replace the whole creation process to get a
`GroupAssignmentRepo` to a simple factory. This increases the speed
of tests and also makes recording some API calls in cassettes
redundant.