Suppress SignalException raised by SIGTERM on deployment
`Puma` raises a `SignalException` on receiving a `SIGTERM` to kill
the process puma was running on. While this is an expected behavior,
deployments, especially for dockerized applications, usually use
`SIGTERM` as a measure to kill processes. This results in our
exception tracker receiving a lot of `SignalException` errors on every
deploy. This small configuration change will suppress those errors.
For more information please check this link:
https://github.com/puma/puma/pull/1690