In progress method starting wtih orgs
We attepted to start with orgs as there were fewer orgs than users and this seemed like a way to limit the amount of api calls. However, in order to check if someone is an admin on an org, we have to make an api call to that org authenticated as someone who is already an admin. That would require going through users in that org to see if one is an admin before running through the remaining users to see if they should be removed or not. It seems simpler now to start with a user and iterate through their orgs (removing them from any they're not an admin for).