Implement allowlist for failbot request params
Closes https://github.com/github/classroom/issues/66

Also, don't send remote_ip or url fields to failbot, so this also closes
https://github.com/github/classroom/issues/67

Similar to https://github.com/github/education-web/pull/2672, this sets
values to '[FILTERED]' in the `params` failbot attribute unless the
param key is in the allowlist.

The idea is that we don't want to accidentally leak sensitive data to
Sentry. If/when we come across some request data that is useful for
debugging in sentry (and also not sensitive), we can add those
attributes to the allowlist.